Added http util methos for code, response and message response
