+++ src/controller/controller.js
@@ -1,4 +1,5 @@
 import validateRequest from '../requestValidation/requestValidation';
+import { code, messageResponse } from '../utils/http';
 
 const { debug, warn, error } = console;
 
@@ -16,7 +17,7 @@ export default (requestSchema, lambdaFunc) => async (event, context) => {
     const body = event.body && JSON.parse(event.body);
     if (process.env.REQUEST_SCHEMA && !requestSchema) {
       error(`[500] The request schema was not found for path: ${path}, schema: ${process.env.REQUEST_SCHEMA}`);
-      return response({ message: '[500] Internal Server Error' }, code.INTERNAL_SERVER_ERROR);
+      return messageResponse('[500] Internal Server Error', code.INTERNAL_SERVER_ERROR);
     }
     if (requestSchema) {
       const { valid, validationResponse } = validateRequest(requestSchema, body);
@@ -34,26 +35,6 @@ export default (requestSchema, lambdaFunc) => async (event, context) => {
   } catch (e) {
     context?.serverlessSdk?.captureError(e);
     error(`[500] Controller failed to execute with a exception for path ${path}, principal: ${principal}`, e);
-    return response({ message: '[500] Internal Server Error' }, code.INTERNAL_SERVER_ERROR);
+    return messageResponse('[500] Internal Server Error', code.INTERNAL_SERVER_ERROR);
   }
-};
-
-export const response = (body, statusCode = code.OK) => {
-  return {
-    statusCode: statusCode,
-    body: JSON.stringify(body),
-    headers: {
-      'Access-Control-Allow-Origin': '*'
-    }
-  };
-};
-
-export const code = {
-  OK: 200,
-  CREATED: 201,
-  BAD_REQUEST: 400,
-  UNAUTHORISED: 401,
-  FORBIDDEN: 403,
-  NOT_FOUND: 404,
-  INTERNAL_SERVER_ERROR: 500
 };
